requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# AI Agent (CrewAI + Groq) - Compatible Python 3.9+
crewai==0.1.32
groq==0.11.0

# Web Scraping
firecrawl-py==4.6.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Google APIs
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0

# Email
email-validator==2.2.0

# Scheduling
APScheduler==3.10.4

# Utilities
python-dateutil==2.9.0
pytz==2024.1

# Logging and monitoring
colorlog==6.8.2

# Testing (optional, for development)
pytest==8.3.3
pytest-mock==3.14.0
pytest-cov==5.0.0
responses==0.25.3
